third_party/blink/web_tests/http/tests/credentialmanager/register-then-sign.html

<!DOCTYPE html>
<title>Credential Manager: end-to-end test that registers/gets an assertion for a credential.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom-lite.js"></script>
<script src="/gen/url/mojom/url.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/authenticator.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/virtual_authenticator.mojom-lite.js"></script>
<script src="resources/test-inputs.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<script>

if (document.location.hostname != "subdomain.example.test") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/register-then-sign.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

promise_test(async _ => {
  // Expect a clean environment.
  let authenticators = await navigator.credentials.test.authenticators();
  assert_equals(authenticators.length, 0);
}, "Set up the testing environment.");

promise_test(async _ => {
  let testAuthenticator = await navigator.credentials.test.createAuthenticator();
  let credential = await navigator.credentials.create({publicKey: MAKE_CREDENTIAL_OPTIONS});
  let getCredentialsOptions = deepCopy(GET_CREDENTIAL_OPTIONS);
  getCredentialsOptions.allowCredentials[0].id = credential.rawId;
  return navigator.credentials.get({publicKey: getCredentialsOptions});
});

promise_test(_ => navigator.credentials.test.clearAuthenticators(), "Clean up testing environment.");

</script>
